// source/frontend/src/models/api/generated/ApiGen_Concepts_FinancialCodeTypes.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
export enum ApiGen_Concepts_FinancialCodeTypes {
  BusinessFunction = 'BusinessFunction',
  CostType = 'CostType',
  WorkActivity = 'WorkActivity',
  ChartOfAccounts = 'ChartOfAccounts',
  FinancialActivity = 'FinancialActivity',
  Responsibility = 'Responsibility',
  YearlyFinancial = 'YearlyFinancial',
}